requirements/requirements.txt
pypdf>=4.1.0
python-docx==0.8.11
watchdog==3.0.0
deepdiff==6.3.0
openai>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
cryptography>=41.0.0
requests>=2.31.0
redis>=5.0.0
pytest>=7.0.0
coverage>=7.0.0
aioitertools>=0.11.0
pywin32>=300; sys_platform == 'win32'
markdown>=3.5.0
pygments>=2.17.0
PyQt6>=6.5.0
qasync>=0.24.0
